requirements.txt
telethon
gspread
google-auth-oauthlib
python-dotenv
rich
requests
google-generativeai>=0.4.0
supabase>=1.0.0
notion-client
schedule 